app/requirements.txt
flask>=3.0,<4.0
pycryptodome>=3.20,<4.0
